Split NumericalEarthCDSAPIExt into a directory extension
Move the flat extension file to ext/NumericalEarthCDSAPIExt/ with one file
per dataset (ERA5, GloFAS, Copernicus land albedo) plus a shared file for
the CDS plumbing they have in common. Definitions move verbatim.
